test des entités + readme + début test contoller

+++ src/Controller/TaskController.php
@@ -9,13 +9,12 @@ use Knp\Component\Pager\PaginatorInterface;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Sensio\Bundle\FrameworkExtraBundle\Configuration\Security;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 
 class TaskController extends AbstractController
 {
-    /**
-     * @Route("/tasks", name="task_list")
-     */
+    #[Route('/tasks', name: 'task_list')]
     public function listAction(EntityManagerInterface $em, Request $request, PaginatorInterface $paginator): Response
     {
         $repo = $em->getRepository(Task::class);
@@ -29,9 +28,35 @@ class TaskController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/tasks/create", name="task_create")
-     */
+    #[Route('/tasks/todo', name: 'task_list_todo')]
+    public function listTodo(EntityManagerInterface $em, Request $request, PaginatorInterface $paginator): Response
+    {
+        $repo = $em->getRepository(Task::class);
+        $task = $repo->findBy(['isDone' => false], ['createdAt' => 'DESC']);
+
+        $taskAll = $paginator->paginate($task, $request->query->getInt('page', 1), 6);
+
+        return $this->render('task/list.html.twig', [
+            'controller_name' => 'TaskController',
+            'tasks' => $taskAll
+        ]);
+    }
+
+    #[Route('/tasks/isDone', name: 'task_list_isDone')]
+    public function listIsDone(EntityManagerInterface $em, Request $request, PaginatorInterface $paginator): Response
+    {
+        $repo = $em->getRepository(Task::class);
+        $task = $repo->findBy(['isDone' => true], ['createdAt' => 'DESC']);
+
+        $taskAll = $paginator->paginate($task, $request->query->getInt('page', 1), 6);
+
+        return $this->render('task/list.html.twig', [
+            'controller_name' => 'TaskController',
+            'tasks' => $taskAll
+        ]);
+    }
+
+    #[Route('/tasks/create', name: 'task_create')]
     public function createAction(Request $request, EntityManagerInterface $em)
     {
 
@@ -62,19 +87,24 @@ class TaskController extends AbstractController
         return $this->render('task/create.html.twig', ['form' => $form->createView()]);
     }
 
-    /**
-     * @Route("/tasks/{id}/edit", name="task_edit")
-     */
+    #[Route('/tasks/{id}/edit', name: 'task_edit')]
     public function editAction(Task $task, Request $request, EntityManagerInterface $em)
     {
+        $user = $this->getUser();
         $form = $this->createForm(TaskType::class, $task);
 
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
-            $em->flush();
+            //dump($task->getUser());
+            //dd($user);
+            if ($user === $task->getUser() || (($task->getUser()->getUsername() === "anonymous") && ($user->getRoles()[0] === "ROLE_ADMIN"))) {
 
-            $this->addFlash('success', 'La tâche a bien été modifiée.');
+                $em->flush();
+                $this->addFlash('success', 'La tâche a bien été modifiée.');
+            } else {
+                $this->addFlash('error', 'Vous n\'avez pas les droits de modifier cette tâche.');
+            }
 
             return $this->redirectToRoute('task_list');
         }
@@ -85,9 +115,7 @@ class TaskController extends AbstractController
         ]);
     }
 
-    /**
-     * @Route("/tasks/{id}/toggle", name="task_toggle")
-     */
+    #[Route('/tasks/{id}/toggle', name: 'task_toggle')]
     public function toggleTaskAction(Task $task, EntityManagerInterface $em)
     {
         $task->toggle(!$task->IsisDone());
@@ -98,15 +126,22 @@ class TaskController extends AbstractController
         return $this->redirectToRoute('task_list');
     }
 
-    /**
-     * @Route("/tasks/{id}/delete", name="task_delete")
-     */
+    #[Route('/tasks/{id}/delete', name: 'task_delete')]
     public function deleteTaskAction(Task $task, EntityManagerInterface $em)
     {
-        $em->remove($task);
-        $em->flush();
+        $user = $this->getUser();
 
-        $this->addFlash('success', 'La tâche a bien été supprimée.');
+        if ($user === $task->getUser() || (($task->getUser()->getUsername() === "anonymous") && ($user->getRoles()[0] === "ROLE_ADMIN"))) {
+
+            $em->remove($task);
+            $em->flush();
+
+            $this->addFlash('success', 'La tâche a bien été supprimée.');
+        } else {
+
+            $this->addFlash('error', 'Vous n\'avez pas les droits pour supprimer cette tâche');
+        }
+
 
         return $this->redirectToRoute('task_list');
     }
